+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Documentation for rocPRIM is available at
 ### Deprecations
 
 * The internal header `detail/match_result_type.hpp` has been deprecated.
+* `TwiddleIn` and `TwiddleOut` have been deprecated in favor of `radix_key_codec`.
 
 ## Unreleased rocPRIM-3.1.0 for ROCm 6.1.0